Tt/bug fix1

+++ src/components/TicketList.js
@@ -4,18 +4,14 @@ import axios from "../axios";
 import "./TicketList.css";
 import Button from "react-bootstrap/Button";
 import Card from "react-bootstrap/Card";
-//import CardGroup from "react-bootstrap/CardGroup";
-import Modal from "react-bootstrap/Modal";
 import Col from "react-bootstrap/Col";
 import Row from "react-bootstrap/Row";
-import Form from "react-bootstrap/Form";
 import Container from "react-bootstrap/Container";
-import Stack from "react-bootstrap/Stack";
 import Dropdown from "react-bootstrap/Dropdown";
-import DropdownButton from "react-bootstrap/DropdownButton";
 import NavBar from "./Nav";
 import TicketModal from "./Modal";
 import formatDate from "../util";
+import { toast } from "react-toastify";
 
 const progressOptions = ["In Progress", "Blocked", "Resolved"];
 function TicketList() {
@@ -29,7 +25,6 @@ function TicketList() {
 
   useEffect(() => {
     async function getTickets() {
-      // console.log("inside getTiclets");
       const tokenData = localStorage.getItem("LoginToken");
       const tokenString = JSON.parse(tokenData);
       const token = tokenString.token;
@@ -51,9 +46,8 @@ function TicketList() {
   useEffect(() => {
     async function getAllStaff() {
       const request = await axios.get("users");
-      // console.log(request.data.data.users);
+
       if (request.data.data.users) {
-        // setTicketData({ staffMembers: request.data.data.users });
         setTicketData((prevState) => ({
           ...prevState,
           staffMembers: request.data.data.users,
@@ -64,7 +58,6 @@ function TicketList() {
   }, []);
 
   const handleSelectProgressDropdown = (ticketId) => async (e) => {
-    // console.log(ticketId);
     const tokenData = localStorage.getItem("LoginToken");
     const tokenString = JSON.parse(tokenData);
     const token = tokenString.token;
@@ -73,16 +66,20 @@ function TicketList() {
       { ticketState: e },
       { headers: { Authorization: `Bearer ${token}` } }
     );
-    // console.log(response);
+
     if (response.status === 200) {
       updateTicketState(response.data.data);
+      toast.success("Ticket state changed !!", {
+        autoClose: 2000,
+        position: "top-right",
+        theme: "colored",
+      });
     }
   };
 
   const updateTicketState = (data) => {
     const newTicketState = data.ticket.ticketState;
     const ticketId = data.ticket._id;
-    // console.log(newTicketState, ticketId);
 
     setTicketData((prevState) => ({
       ...prevState,
@@ -96,8 +93,6 @@ function TicketList() {
           : ticket
       ),
     }));
-
-    // console.log(ticketData);
   };
 
   const handleCardClick = (e, i) => {
@@ -119,31 +114,38 @@ function TicketList() {
 
   async function handleDeleteTicket(e, i) {
     e.preventDefault();
-    const currentTicket = ticketData.data[i];
-    const tokenData = localStorage.getItem("LoginToken");
-    const tokenString = JSON.parse(tokenData);
-    const token = tokenString.token;
-    await axios.delete(`tickets/${currentTicket._id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    setTicketDeleted(true);
+    if (window.confirm("Are you sure you want to delete")) {
+      const currentTicket = ticketData.data[i];
+      const tokenData = localStorage.getItem("LoginToken");
+      const tokenString = JSON.parse(tokenData);
+      const token = tokenString.token;
+      const response = await axios.delete(`tickets/${currentTicket._id}`, {
+        headers: { Authorization: `Bearer ${token}` },
+      });
+
+      if (response.status === 204) {
+        setTicketDeleted(true);
+        toast.success("ticket deleted", {
+          autoClose: 2000,
+        });
+      }
+
+      if (ticketDeleted === true) setTicketDeleted(false);
+    }
   }
 
   const updateCommentCallback = (ticketId, incommingComment) => {
     const newComment = incommingComment;
-    // console.log(newComment, "newcomment");
+    const updatedData = ticketData.data.map((ticket) =>
+      ticket._id === ticketId
+        ? { ...ticket, comments: [...ticket.comments, newComment] }
+        : ticket
+    );
+    const index = ticketData.data.findIndex(({ _id }) => _id === ticketId);
+    setModalData(updatedData[index]);
     setTicketData((prevState) => ({
       ...prevState,
-      data: prevState.data.map((ticket) => {
-        if (ticket._id === ticketId) {
-          const updatedTicket = {
-            ...ticket,
-            comments: [...ticket.comments, newComment],
-          };
-          setModalData(updatedTicket);
-        }
-        return ticket;
-      }),
+      data: updatedData,
     }));
   };
 
@@ -193,7 +195,10 @@ function TicketList() {
                       <Card.Subtitle className="mb-2 text-muted">
                         {ticket.createdBy}
                       </Card.Subtitle>
-                      <Card.Text>{ticket.description}</Card.Text>
+                      <Card.Text>{`${ticket.description.substring(
+                        0,
+                        30
+                      )}...`}</Card.Text>
                       <div
                         style={{
                           display: "flex",
@@ -258,18 +263,3 @@ function TicketList() {
 }
 
 export default TicketList;
-// const assignTicketCallback = ({ ticketId, ...ticketAssignedDetails }) => {
-
-//   setTicketData((prevState) => ({
-//     ...prevState,
-//     data: prevState.data.map((ticket) => {
-//       console.log(ticket._id, ticketId);
-//       return ticket._id === ticketId
-//         ? {
-//             ...ticket,
-//             assignedTo: ticketAssignedDetails,
-//           }
-//         : ticket;
-//     }),
-//   }));
-// };
